+++ Utils.lua
@@ -33,7 +33,18 @@ function Private.Utils.RollDice()
 end
 
 function Private.Utils.FindThirdPartyGroupFrameForUnit(unit, kind)
-	if DandersFrames and DandersFrames.Api and DandersFrames.Api.GetFrameForUnit then
+	if Grid2 and Grid2LayoutFrame and Grid2LayoutHeader1 then
+		for i = 1, 5 do
+			local name = string.format("Grid2LayoutHeader1UnitButton%d", i)
+			local maybeFrame = _G[name]
+
+			if maybeFrame and maybeFrame.unit == unit then
+				return maybeFrame
+			end
+		end
+
+		return nil
+	elseif DandersFrames and DandersFrames.Api and DandersFrames.Api.GetFrameForUnit then
 		local frame = DandersFrames.Api.GetFrameForUnit(unit, kind)
 
 		if frame then